services/youtube-research/requirements.txt
youtube-transcript-api
yt-dlp
openai
httpx
python-dotenv
